- Move ui-router bindings to a single source file - Remove the leading "I" of some traits.
